Minor clean up, revert makefile change, include submodules directly in rcore
